controllers y rutas para compra de VideoGame y carga de saldo

+++ src/controllers/Users/deleteFavorites.js
@@ -1,7 +1,9 @@
 const{Users, Videogames, Favorites} = require('../../db')
 
 const deleteFavorite= async (idUser, idVideogames)=>{
-    const result = await Favorites.destroy({
+      const gameFav = await Favorites.findOne({where:{VideogameId: idVideogames, UserId: idUser}})
+  if(gameFav.buy ===true) throw new Error('este juego esta comprado, no se puede eliminar')
+      const result = await Favorites.destroy({
         where: { VideogameId: idVideogames, UserId: idUser }
       });
